+++ include/Gaffer/Context.h
@@ -37,7 +37,6 @@
 #ifndef GAFFER_CONTEXT_H
 #define GAFFER_CONTEXT_H
 
-#include "Gaffer/CatchingSignalCombiner.h"
 #include "Gaffer/Export.h"
 #include "Gaffer/Signals.h"
 #include "Gaffer/ThreadState.h"
@@ -93,7 +92,7 @@ class GAFFER_API Context : public IECore::RefCounted
 
 		IE_CORE_DECLAREMEMBERPTR( Context )
 
-		using ChangedSignal = Signals::Signal<void ( const Context *context, const IECore::InternedString & ), CatchingSignalCombiner<void>>;
+		using ChangedSignal = Signals::Signal<void ( const Context *context, const IECore::InternedString & ), Signals::CatchingCombiner<void>>;
 
 		/// Sets a variable to the specified value. A copy is taken so that
 		/// subsequent changes to `value` do not affect the context.