+++ include/GafferDispatch/Dispatcher.h
@@ -40,7 +40,6 @@
 #include "GafferDispatch/Export.h"
 #include "GafferDispatch/TaskNode.h"
 
-#include "Gaffer/CatchingSignalCombiner.h"
 #include "Gaffer/NumericPlug.h"
 #include "Gaffer/Signals.h"
 
@@ -107,8 +106,8 @@ class GAFFERDISPATCH_API Dispatcher : public Gaffer::Node
 		GAFFER_NODE_DECLARE_TYPE( GafferDispatch::Dispatcher, DispatcherTypeId, Gaffer::Node );
 
 		using PreDispatchSignal = Gaffer::Signals::Signal<bool (const Dispatcher *, const std::vector<TaskNodePtr> &), Detail::PreDispatchSignalCombiner>;
-		using DispatchSignal = Gaffer::Signals::Signal<void (const Dispatcher *, const std::vector<TaskNodePtr> &), Gaffer::CatchingSignalCombiner<void>>;
-		using PostDispatchSignal = Gaffer::Signals::Signal<void (const Dispatcher *, const std::vector<TaskNodePtr> &, bool), Gaffer::CatchingSignalCombiner<void>>;
+		using DispatchSignal = Gaffer::Signals::Signal<void (const Dispatcher *, const std::vector<TaskNodePtr> &), Gaffer::Signals::CatchingCombiner<void>>;
+		using PostDispatchSignal = Gaffer::Signals::Signal<void (const Dispatcher *, const std::vector<TaskNodePtr> &, bool), Gaffer::Signals::CatchingCombiner<void>>;
 		//! @name Dispatch Signals
 		/// These signals are emitted on dispatch events for any registered Dispatcher instance.
 		////////////////////////////////////////////////////////////////////////////////////////