+++ Changes.md
@@ -16,6 +16,7 @@ Breaking Changes
     - Boost `snake_case` naming has been replaced with `CamelCase`.
     - Connection groups are not supported. Use `connectFront()` to connect in front of existing slots.
   - Removed the `Gaffer/BlockedConnection.h` header file. BlockedConnection can now be found in the Signals namespace provided by `Gaffer/Signals.h`.
+  - Remove the `Gaffer/CatchingSignalCombiner.h` header file. CatchingSignalCombiner can now be found as `Signals::CatchingCombiner` in `Gaffer/Signals.h`.
   - Moved all Python classes into the `Gaffer.Signals` submodule.
 - Replaced all usage of `boost::optional` with `std::optional`.